+++ packages/obsidian-plugin/src/main.ts
@@ -1,5 +1,6 @@
 import { Notice, Plugin, type TAbstractFile, type WorkspaceLeaf } from "obsidian";
 import { ReviewService } from "../../core/src/service/review-service";
+import { userDataReviewStorageBase } from "../../core/src/storage/user-data";
 import { ReviewWatcher } from "./watcher/review-watcher";
 import { ReviewGateView, REVIEW_GATE_VIEW_TYPE } from "./ui/review-view";
 
@@ -15,7 +16,8 @@ export default class ObsidianReviewGatePlugin extends Plugin {
       return;
     }
     const vaultRoot = getBasePath.call(this.app.vault.adapter);
-    const opened = await ReviewService.open(vaultRoot);
+    const storageBase = userDataReviewStorageBase(vaultRoot);
+    const opened = await ReviewService.open(vaultRoot, { storageBase });
     this.service = opened.service;
 
     this.registerView(
@@ -34,7 +36,7 @@ export default class ObsidianReviewGatePlugin extends Plugin {
       callback: () => void this.refreshViews(),
     });
 
-    this.watcher = new ReviewWatcher(vaultRoot, async () => {
+    this.watcher = new ReviewWatcher(storageBase, async () => {
       await this.refreshViews();
     });
     await this.watcher.start();